backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pandas==2.1.4
openpyxl==3.1.2
fake-useragent==1.4.0
asyncio==3.4.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4